Move four more Enterprise-only controllers into pkg/enterprise
Manager, non-cluster host, authentication and policy recommendation.
