+++ pkg/enterprise/controller/manager/manager_controller_cloud_test.go
