+++ pkg/enterprise/controller/manager/manager_controller_suite_test.go
@@ -29,6 +29,6 @@ func TestStatus(t *testing.T) {
 	logf.SetLogger(zap.New(zap.WriteTo(ginkgo.GinkgoWriter), zap.UseDevMode(true), zap.Level(uzap.NewAtomicLevelAt(uzap.DebugLevel))))
 	gomega.RegisterFailHandler(ginkgo.Fail)
 	suiteConfig, reporterConfig := ginkgo.GinkgoConfiguration()
-	reporterConfig.JUnitReport = "../../../report/ut/manager_controller_suite.xml"
-	ginkgo.RunSpecs(t, "pkg/controller/manager Suite", suiteConfig, reporterConfig)
+	reporterConfig.JUnitReport = "../../../../report/ut/manager_controller_suite.xml"
+	ginkgo.RunSpecs(t, "pkg/enterprise/controller/manager Suite", suiteConfig, reporterConfig)
 }
