+++ pkg/enterprise/controller/manager/manager_controller_test.go
