+++ pkg/enterprise/controller/nonclusterhost/nonclusterhost_controller_test.go
