+++ pkg/enterprise/controller/policyrecommendation/policyrecommendation_suite_test.go
@@ -1,10 +1,10 @@
-// Copyright (c) 2024-2026 Tigera, Inc. All rights reserved.
+// Copyright (c) 2023-2026 Tigera, Inc. All rights reserved.
 
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
 //
-//     http://www.apache.org/licenses/LICENSE-2.0
+//      http://www.apache.org/licenses/LICENSE-2.0
 //
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package nonclusterhost
+package policyrecommendation
 
 import (
 	"testing"
@@ -29,6 +29,6 @@ func TestStatus(t *testing.T) {
 	logf.SetLogger(zap.New(zap.WriteTo(ginkgo.GinkgoWriter), zap.UseDevMode(true), zap.Level(uzap.NewAtomicLevelAt(uzap.DebugLevel))))
 	gomega.RegisterFailHandler(ginkgo.Fail)
 	suiteConfig, reporterConfig := ginkgo.GinkgoConfiguration()
-	reporterConfig.JUnitReport = "../../../report/ut/nonclusterhost_controller_suite.xml"
-	ginkgo.RunSpecs(t, "pkg/controller/nonclusterhost Suite", suiteConfig, reporterConfig)
+	reporterConfig.JUnitReport = "../../../../report/ut/policyrecommendation_controller_suite.xml"
+	ginkgo.RunSpecs(t, "pkg/enterprise/controller/policyrecommendation Controller Suite", suiteConfig, reporterConfig)
 }
